Fixes https://github.com/dgraph-io/badger/issues/1263 

This PR proposes use of https://github.com/FastFilter/xorfilter instead of `bbloom` https://github.com/dgraph-io/ristretto/blob/master/z/bbloom.go .

Following are the performance benchmarks
```
name                           time/op
Set/bloom-8              23.2ms ± 3%
Set/xorFilter-8           121ms ± 4%
Get/success/bloom-8      20.9ns ± 5%
Get/success/xorFilter-8  7.41ns ± 7%
Get/failure/bloom-8      8.27ns ± 6%
Get/failure/xorFilter-8  7.52ns ± 5%
```
The size xor filter is lesser than bbloom.
```
Number of keys = 2070594 (~ 2 million)
Size of bloom filter  	= 5592436 (5.59 MB)
Size of xor filter	= 3395912 (3.39 MB)
```

Xorfilter is definitely faster than bbloom filter but there isn't a significant performance difference in `key iterator` or `get` operation.
```
name           old time/op  new time/op  delta
KeyIterator-8   123µs ± 7%   121µs ± 7%   ~     (p=0.436 n=10+10)
```

Benchmarks are stored in `builder_test.go` and `iterator_test.go`

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/badger/1264)
<!-- Reviewable:end -->