@jarifibrahim 2020-03-24
This is a breaking change and the performance gain is not worth breaking the format. 
The bloom filters are used only to pick the tables which happens when `NewKeyIterator` is called or `txn.Get` is called. The change doesn't have any real performance gains.